🚚 Renamed services to assets
