+++ zabbixci/assets/__init__.py
