+++ zabbixci/assets/asset.py
@@ -1,0 +1,5 @@
+from abc import ABCMeta
+
+
+class Asset(object, metaclass=ABCMeta):
+    pass
